requirements.txt
anthropic>=0.40.0
matplotlib>=3.8.0
rich>=13.0.0
python-dotenv>=1.0.0
